requirements.txt
telethon==1.38.1
python-dotenv==1.0.1
langgraph==0.2.63
langchain_openai==0.3.0
langchain==0.3.14
pytz==2024.2